Query.SP: replace 'Real' with 'EdgeWeight' type class
The 'Real' type class (including 'Num') defines 7 functions,
out of which only two are used by the shortest path algorithm
((+) and fromInteger). This is replaced by 'EdgeWeight', which
more accurately describes the required properties of the edge
labels for the shortest path algorithm ('Ord' and 'Monoid').

The 'EdgeWeight' type class describes the operations
needed for Dijkstra:
* comparing two edge weights ('Ord')
* combining two edge weights (Monoid.mappend)
* the absence of an edge weight (Monoid.mempty)

Usually, edge weights are combined by summing and the
absence of an edge weight is zero.
But other ways to combine edge weights are also useful.
For example, in a graph wheres currencies are nodes,
and edge weights are exchange rates -- and we want to find
the lowest price for which we can purchase a currency --
we'd want "combine" to be multiplication and "nothing" to
be equal to 1.